// src/app/shared/models/patient.ts
export class Patient {
    id: number;
    name: string;
    dob: string;
    age: number;
    careCoordinatorId: number;
    zoningCategory: string;
    alert1: string;
    alert1SD: string;
    alert1ED: string;
    alert2: string;
    alert2SD: string;
    alert2ED: string;
    alert3: string;
    alert3SD: string;
    alert3ED: string;
  }
  